requirements.txt
Django==1.9.5
django-bootstrap3==7.0.1
django-datetime-widget==0.9.3
numpy==1.11.0
psycopg2==2.6.1
pytz==2016.3
Pillow==2.9.0
